// angularRouting/src/app/product/models/product.ts
export class Product {
    Id: number;
      SubTypeCode: string;
      CompanyId: number;
      StatusCode: string;
      BussinessLevelCode: string;
      UserId: number;
      Description: string;
      Name: string;
      MediaId: string;
      MediaPath: string;
      UserdOnCompanyStore: boolean;
      ProductCount: number;
      WebsiteCout: number;
      DisplaySequence: number;
      IsVisible: boolean;
    }